requirements.txt
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.5.14
deeplake==3.9.45
dill==0.4.0
distro==1.9.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.2
fsspec==2025.5.1
gcloud==0.18.3
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0
groq==0.26.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humbug==0.3.2
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.65
langchain-google-community==2.0.7
langchain-google-genai==2.1.5
langchain-groq==0.3.2
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
libdeeplake==0.0.147
lz4==4.4.4
marshmallow==3.26.1
multidict==6.4.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numexpr==2.11.0
numpy==1.26.4
oauth2client==4.1.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pathos==0.3.4
pillow==10.4.0
pox==0.3.6
ppft==1.7.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.11.3
selenium==4.15.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
tenacity==8.5.0
tiktoken==0.4.0
tokenizers==0.21.1
tqdm==4.66.6
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
